backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9        # required for file uploads in FastAPI

# ── Auth ────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Database ────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0              # async SQLite driver for local dev

# ── ML core ────────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# ── ML classifiers ─────────────────────────────────────────────────────
scikit-learn>=1.4.0
joblib>=1.3.0
xgboost>=2.0.0
lightgbm>=4.3.0

# ── Feature extraction ─────────────────────────────────────────────────
scipy>=1.11.0                  # used indirectly by some sklearn estimators

# ── PDF report generation ──────────────────────────────────────────────
reportlab>=4.1.0

# ── Utilities ───────────────────────────────────────────────────────────
python-dotenv==1.0.1           # loads .env file
pydantic[email]==2.7.1         # request/response validation
pydantic-settings==2.2.1
